# Optimize Sprintfln() by used fixed allocate slice instead of append